// packages/shared/src/database.types.d.ts
export type Database = {
    __InternalSupabase: {
        PostgrestVersion: "13.0.4";
    };
    public: {
        Tables: {
// ...
            profiles: {
                Row: {
                    avatar_url: string | null;
                    beta_reader_status: Database["public"]["Enums"]["beta_reader_status"];
                    created_at: string;
                    display_name: string | null;
                    id: string;
                    role: Database["public"]["Enums"]["user_role"];
                    updated_at: string;
                    username: string | null | null;
                    website: string | null;
                };
// ...
            };
// ...
        };
// ...
        Enums: {
            beta_application_status: "pending" | "approved" | "denied";
            beta_reader_status: "not_applied" | "pending" | "approved" | "rejected";
            content_block_type: "heading_1" | "heading_2" | "heading_3" | "paragraph" | "bullet_list" | "ordered_list" | "image" | "table" | "quote" | "code" | "divider";
            discount_type: "percentage" | "fixed_amount";
            product_type: "single_issue" | "bundle" | "chapter_pass" | "arc_pass";
            subscription_status: "incomplete" | "incomplete_expired" | "trialing" | "active" | "past_due" | "canceled" | "unpaid" | "paused";
            user_role: "admin" | "support" | "accountant" | "user" | "super_admin";
            work_status: "planning" | "writing" | "editing" | "published" | "on_hold";
            work_type: "book" | "volume" | "saga" | "arc" | "issue";
        };
        CompositeTypes: {
            [_ in never]: never;
        };
    };
};
type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;
type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">];
// ...
export type Enums<DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"] | {
    schema: keyof DatabaseWithoutInternals;
}, EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
} ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"] : never = never> = DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
} ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName] : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"] ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions] : never;